server/requirements.txt
beautifulsoup4==4.11.2
Flask==2.2.3
Flask_Cors==3.0.10
numpy==1.24.1
pandas==1.5.3
openai==0.27.4
transformers==4.29.0
Pillow==9.5.0
